Renamed and removed vortex code
