--- src/main/java/app/ar/com/dgarcia/processing/sandbox/vortex/VortexEndpoint.java
@@ -1,8 +1,0 @@
-package app.ar.com.dgarcia.processing.sandbox.vortex;
-
-/**
- * This type represents a communication endpoint (can be half or full duplex)
- * Created by ikari on 17/01/2015.
- */
-public interface VortexEndpoint {
-}
