--- src/main/java/app/ar/com/dgarcia/processing/sandbox/vortex/VortexStream.java
@@ -1,8 +1,0 @@
-package app.ar.com.dgarcia.processing.sandbox.vortex;
-
-/**
- * Created by kfgodel on 18/01/15.
- */
-public interface VortexStream {
-    void receive(Object message);
-}
