--- src/main/java/app/ar/com/dgarcia/processing/sandbox/vortex/VortexTunnel.java
@@ -1,9 +1,0 @@
-package app.ar.com.dgarcia.processing.sandbox.vortex;
-
-/**
- * Created by ikari on 17/01/2015.
- */
-public interface VortexTunnel {
-
-    void receive(Object message);
-}
