+++ src/main/java/ar/com/kfgodel/vortex/api/manifest/ReceiverManifest.java
@@ -1,0 +1,20 @@
+package ar.com.kfgodel.vortex.api.manifest;
+
+import java.util.function.Consumer;
+
+/**
+ * This type represents the definition of a consumer expectation and obligations towards a VortexNode
+ * Created by kfgodel on 18/01/15.
+ */
+public interface ReceiverManifest {
+
+    Consumer<Object> onAvailableProducers();
+
+    void onNoAvailableProducers();
+
+    boolean isCompatibleWith(EmitterManifest emitterManifest);
+
+    VortexInterest getInterest();
+    void changeInterest(VortexInterest newInterest);
+    void setInterestChangeListener(Runnable changeListener);
+}
