+++ src/main/java/ar/com/kfgodel/vortex/impl/manifest/AllInterest.java
@@ -1,0 +1,18 @@
+package ar.com.kfgodel.vortex.impl.manifest;
+
+import ar.com.kfgodel.vortex.api.manifest.VortexInterest;
+
+/**
+ * This type represents the interest that accepts anything (all types of messages)
+ * Created by ikari on 21/01/2015.
+ */
+public class AllInterest implements VortexInterest {
+
+    public static AllInterest INSTANCE = new AllInterest();
+
+    @Override
+    public boolean intersects(VortexInterest otherInterest) {
+        // Interseccion con vacio da vacio
+        return true && !otherInterest.equals(NoInterest.INSTANCE);
+    }
+}
